additionalFiles/streaming/algorithimsCode/requirements.txt
certifi==2020.12.5
chardet==3.0.4
idna==2.10
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.26.2
Pillow==7.0.0

